requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.6
python-dotenv==1.0.1
pydantic==2.10.6
requests==2.32.3
aiohttp==3.11.12
pyngrok==7.2.3
fastapi-redis-rate-limiter==1.0.1
APScheduler==3.11.0
pytz==2021.3
PyJWT==2.10.1
cryptography==42.0.5